tutorials/03-advanced/image_captioning/requirements.txt
matplotlib
nltk
numpy
Pillow
argparse